+++ .github/CHANGELOG.md
@@ -132,6 +132,7 @@
 1. [A32NX/FWS] Added Five Hundred callout only not on glide option - @BravoMike99 (bruno_pt99)
 1. [A32NX] Split navigation lights into two separate systems - @tracernz (Mike)
 1. [A380X/FG] Fixed autopilot managed speed target not limited by characteristic speeds outside of approach - @BravoMike99 (bruno_pt99)
+1. [FWS] Add Avionics Test Mode pin program to display flight phase number on EWD - @FozzieHi (fozzie)
 
 ## 0.14.0